pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
